pyproject.toml
[project]
name = "llm-evals"
version = "0.1.0"
description = "Config-driven LLM evaluation framework using OpenRouter"
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "openai>=1.0",
    "pandas>=2.0",
    "python-dotenv>=1.0",
]

[project.scripts]
llm-evals = "llm_evals.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[tool.uv]
package = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
